Revert "Adds tests"
This reverts commit cac4da4837b3817b74de9a6252dfc987e529c9c7.
