+++ backend/tests/v2/factories/__init__.py
@@ -14,7 +14,6 @@ from tests.v2.factories.pending_snapshots import PendingEpochSnapshotFactorySet
 from tests.v2.factories.finalized_snapshots import FinalizedEpochSnapshotFactorySet
 from tests.v2.factories.patrons import PatronModeEventFactorySet
 from tests.v2.factories.tos import UserConsentsFactorySet
-from tests.v2.factories.rewards import RewardFactorySet
 from dataclasses import dataclass
 
 
@@ -44,7 +43,6 @@ class FactoriesAggregator:
     score_delegations: ScoreDelegationFactorySet
     deposits: DepositFactorySet
     multisig_signatures: MultisigSignatureFactorySet
-    rewards: RewardFactorySet
 
     def __init__(self, fast_session: AsyncSession):
         """
@@ -64,4 +62,3 @@ class FactoriesAggregator:
         self.score_delegations = ScoreDelegationFactorySet(fast_session)
         self.deposits = DepositFactorySet(fast_session)
         self.multisig_signatures = MultisigSignatureFactorySet(fast_session)
-        self.rewards = RewardFactorySet(fast_session)
